Don't use nose for test discovery + add parallel builds

+++ testrig/__init__.py
@@ -1,38 +1,57 @@
 from __future__ import absolute_import, division, print_function
 
 import sys
-import nose
 import os
 import shutil
 import tempfile
 import textwrap
 import subprocess
 import virtualenv
+import pkgutil
 
-CACHE_DIR = os.path.normpath(os.path.join(os.path.abspath(os.path.dirname(__file__)), '..', 'cache'))
-DOWNLOAD_DIR = os.path.join(CACHE_DIR, 'download')
-ENV_DIR = os.path.join(CACHE_DIR, 'env')
-CODE_DIR = os.path.join(CACHE_DIR, 'code')
+class Fixture(object):
+    """
+    Fixture for running test suites.
 
-class TestIntegration(object):
-    TEST_LOG = sys.stderr
+    Methods
+    -------
+    run_cmd
+    run_python_code
+    run_python_script
+    run_pip
+    run_numpytest
+    pip_install
+    git_install
+    get_repo
+    print
+    print_header
+    setup
+    teardown
 
-    def __init__(self):
-        self.tmp_virtualenv = None
+    """
 
-        for d in [DOWNLOAD_DIR, ENV_DIR, CODE_DIR]:
+    def __init__(self, name, cache_dir, log_fn, clean_build=True, log_prefix=False):
+        self.name = name
+        self.log = open(log_fn, 'wb')
+        self.log_fn = log_fn
+        self.log_prefix = log_prefix
+        self.clean_build = clean_build
+        self.download_dir = os.path.join(cache_dir, 'download')
+        self.env_dir = os.path.join(cache_dir, 'env')
+        self.code_dir = os.path.join(cache_dir, 'code')
+
+    def setup(self):
+        for d in (self.download_dir, self.code_dir):
             if not os.path.isdir(d):
                 os.makedirs(d)
 
-    def setup(self):
-        if os.path.isdir(ENV_DIR):
-            shutil.rmtree(ENV_DIR)
-        self.tmp_virtualenv = ENV_DIR
-        virtualenv.create_environment(self.tmp_virtualenv)
+        if os.path.isdir(self.env_dir):
+            shutil.rmtree(self.env_dir)
+
+        virtualenv.create_environment(self.env_dir)
 
     def teardown(self):
-        if self.tmp_virtualenv is not None:
-            shutil.rmtree(self.tmp_virtualenv)
+        pass
 
     def run_python_code(self, code):
         tmpd = os.path.abspath(tempfile.mkdtemp())
@@ -42,14 +61,14 @@ class TestIntegration(object):
             with open(fn, 'w') as f:
                 f.write(code)
             os.chdir(tmpd)
+            self.print(("$ cat <<__EOF__ > %s\n" % fn) + code + "\n__EOF__")
             self.run_python_script([fn])
         finally:
             shutil.rmtree(tmpd)
             os.chdir(cwd)
 
     def run_python_script(self, cmd, **kwargs):
-        assert self.tmp_virtualenv is not None
-        cmd = [os.path.join(self.tmp_virtualenv, 'bin', 'python')] + cmd
+        cmd = [os.path.join(self.env_dir, 'bin', 'python')] + cmd
         self.run_cmd(cmd, **kwargs)
 
     def run_cmd(self, cmd, **kwargs):
@@ -58,23 +77,21 @@ class TestIntegration(object):
             msg = '(cd %r && %s)' % (kwargs['cwd'], msg)
         msg = '$ ' + msg
 
-        print(msg, file=sys.stderr)
-        if self.TEST_LOG is not sys.stderr:
-            print(msg, file=self.TEST_LOG)
+        self.print(msg)
 
-        p = subprocess.Popen(cmd, stdout=self.TEST_LOG, stderr=subprocess.STDOUT,
+        p = subprocess.Popen(cmd, stdout=self.log, stderr=subprocess.STDOUT,
                              **kwargs)
         try:
             p.communicate()
             if p.returncode != 0:
-                raise RuntimeError("Failed to run %r (see log)")
+                raise RuntimeError("Failed to run %r (see log)" % (cmd,))
         except:
-            p.terminate()
+            if p.returncode is None:
+                p.terminate()
             raise
 
     def run_pip(self, cmd):
-        assert self.tmp_virtualenv is not None
-        cmd = [os.path.join(self.tmp_virtualenv, 'bin', 'pip')] + cmd
+        cmd = [os.path.join(self.env_dir, 'bin', 'pip')] + cmd
         return self.run_python_script(cmd)
 
     def run_numpytest(self, module):
@@ -85,17 +102,18 @@ class TestIntegration(object):
         """) % (module,))
 
     def pip_install(self, requirements):
-        with tempfile.NamedTemporaryFile() as f:
-            f.write(requirements)
-            f.flush()
-            self.run_pip(['install', '-r', f.name, '--use-mirrors',
-                          '--download-dir', DOWNLOAD_DIR])
+        parts = [x.strip() for x in requirements.splitlines()
+                 if x.strip()]
+        self.run_pip(['install', '--use-mirrors', '--download-cache', self.download_dir] + parts)
 
-    def git_install(self, srcs):
+    def git_install(self, srcs, setup_py=None):
         for module, src_repo, branch in srcs:
             self._git_install_one(module, src_repo, branch)
 
-    def _git_install_one(self, module, src_repo, branch):
+    def _git_install_one(self, module, src_repo, branch, setup_py=None):
+        if setup_py is None:
+            setup_py = 'setup.py'
+
         repo = self.get_repo(module)
 
         if not os.path.isdir(repo):
@@ -103,33 +121,89 @@ class TestIntegration(object):
 
         self.run_cmd(['git', 'fetch', 'origin'], cwd=repo)
         self.run_cmd(['git', 'reset', '--hard', 'origin/' + branch], cwd=repo)
-        self.run_cmd(['git', 'clean', '-f', '-d', '-x'], cwd=repo)
+        if self.clean_build:
+            self.run_cmd(['git', 'clean', '-f', '-d', '-x'], cwd=repo)
 
-        self.run_python_script(['setup.py', 'install'], cwd=repo)
+        self.run_python_script([setup_py, 'install'], cwd=repo)
 
     def get_repo(self, module):
-        return os.path.join(CODE_DIR, module)
+        return os.path.join(self.code_dir, module)
 
-    def print_message(self):
+    def print_header(self, name):
         msg = ""
         msg += ("-"*79) + "\n"
-        msg += ("Running test: %s" % self.__class__.__name__) + "\n"
+        msg += ("Running test: %s" % name) + "\n"
         msg += ("-"*79) + "\n"
+        msg += "Logging into: %r\n" % self.log_fn
+        
+        self.print(msg)
 
-        print(msg, file=sys.stderr)
-        if self.TEST_LOG is not sys.stderr:
-            print(msg, file=self.TEST_LOG)
-
-
-def run(modules=None, log=None):
-    if log is not None:
-        if isinstance(log, str):
-            TestIntegration.TEST_LOG = open(log, 'wb')
+    def print(self, msg):
+        if self.log_prefix:
+            prefix = self.name + ": "
+            out_msg = prefix + msg.replace("\n", "\n" + prefix)
+            print(out_msg, file=sys.stderr)
         else:
-            TestIntegration.TEST_LOG = log
+            print(msg, file=sys.stderr)
+        sys.stderr.flush()
+        if self.log is not sys.stderr:
+            print(msg, file=self.log)
+            self.log.flush()
 
-    if modules is None:
-        nose.main(argv=['-vv'])
-    else:
-        targets = ['testrig.test_' + x for x in modules]
-        nose.main(argv=['-vv'] + targets)
+
+def get_tests():
+    import testrig
+    path = testrig.__path__
+    names = []
+    for importer, modname, ispkg in pkgutil.iter_modules(path):
+        if not modname.startswith('test_'):
+            continue
+        name = modname[5:]
+
+        modname = 'testrig.test_' + name
+        mod = __import__(modname, fromlist='x')
+        if mod.__doc__:
+            info = mod.__doc__.strip()
+        else:
+            info = ""
+
+        names.append((name, info))
+
+    return names
+
+
+def run(name, base_dir, clean_build=True, log_prefix=False):
+    modname = 'testrig.test_' + name
+    try:
+        mod = __import__(modname, fromlist='x')
+    except ImportError:
+        print("Test module %r does not exist" % (modname,))
+        return False
+
+    cache_dir = os.path.join(base_dir, name)
+    if not os.path.isdir(cache_dir):
+        os.makedirs(cache_dir)
+
+    log_fn = os.path.join(cache_dir, 'test-%s.log' % name)
+    fixture = Fixture(name,
+                      cache_dir=cache_dir, log_fn=log_fn,
+                      clean_build=clean_build,
+                      log_prefix=log_prefix)
+    fixture.print_header(name)
+    try:
+        fixture.setup()
+        try:
+            mod.run(fixture)
+            fixture.print("OK")
+            return True
+        except Exception as exc:
+            import traceback
+            fixture.print(traceback.format_exc())
+            if isinstance(exc, AssertionError):
+                fixture.print("FAILURE")
+                return False
+            else:
+                fixture.print("ERROR")
+                return False
+    finally:
+        fixture.teardown()
